Move rotate and translate functions into the rotate module
